AGSEvent refactoring- split to 2 files
